backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
PyYAML==6.0.2
Jinja2==3.1.5
python-dotenv==1.0.1
playwright==1.49.1
httpx==0.28.1
pytest==8.3.4
